requirements.txt
huggingface_hub==0.23.4
langchain==0.2.6
python-dotenv==1.0.1
PyYAML==6.0.1
torch==2.3.0
transformers==4.42.3
datasets==2.20.0
pandas==2.2.2
accelerate==0.32.1
bitsandbytes==0.43.1
bleurt @ git+https://github.com/google-research/bleurt.git
